Revert horrible PRIuPTR and change to unsinged long cast style.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/apps/gworkspace/trunk@37352 72102866-910b-0410-8b05-ffd578937521

+++ GWMetadata/MDKit/MDKResultsCategory.m
@@ -296,8 +296,8 @@ static NSImage *whiteArrowDown = nil;
 
       if (range.length > MIN_LINES) {
         str = NSLocalizedString(@"more...", @"");
-        str = [NSString stringWithFormat: @"%"PRIuPTR" %@",
-			([results count] - MIN_LINES), str];
+        str = [NSString stringWithFormat: @"%lu %@",
+			(unsigned long)([results count] - MIN_LINES), str];
         [topFiveFootButt setTitle: str];
         [topFiveFootButt setTag: ALL_RESULTS];      
       }
